# api/openapi.yaml
openapi: 3.0.3
info:
  title: Opal Custom App Connector API
  version: "1.0"
servers:
- url: /
paths:
  /status:
    get:
      description: Returns the status of the custom connector API.
      operationId: getStatus
      parameters:
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: The status of the custom connector API.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - status
  /users:
    get:
      description: Returns a list of users for your custom connector app.
      operationId: getUsers
      parameters:
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
          description: One page of users that exist for your custom app.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - users
  /resources:
    get:
      description: Returns a list of resources for your organization.
      operationId: getResources
      parameters:
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcesResponse'
          description: One page of resources that exist for your custom app.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - resources
  /resources/{resource_id}:
    get:
      description: Get a specific resource by its id.
      operationId: getResource
      parameters:
      - description: The id of the resource to fetch.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: false
        in: path
        name: resource_id
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceResponse'
          description: The resource that was requested.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - resources
  /resources/{resource_id}/access_levels:
    get:
      description: Get the access level list of a resource.
      operationId: getResourceAccessLevels
      parameters:
      - description: The id of the resource
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: false
        in: path
        name: resource_id
        required: true
        schema:
          type: string
        style: simple
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceAccessLevelsResponse'
          description: One page of access levels for the resource in question.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - resources
  /resources/{resource_id}/users:
    get:
      description: Get the user access list of a resource.
      operationId: getResourceUsers
      parameters:
      - description: The id of the resource
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: false
        in: path
        name: resource_id
        required: true
        schema:
          type: string
        style: simple
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceUsersResponse'
          description: One page of users with access to the resource in question.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - resources
    post:
      description: Adds a user to the access list of the specified resource.
      operationId: addResourceUser
      parameters:
      - description: The id of the resource.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: false
        in: path
        name: resource_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addResourceUser_request'
        required: true
      responses:
        "200":
          description: The user was successfully added to the resource.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - resources
  /resources/{resource_id}/users/{user_id}:
    delete:
      description: Remove a user from the access list of the specified resource.
      operationId: removeResourceUser
      parameters:
      - description: The id of the resource.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: false
        in: path
        name: resource_id
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the user.
        example: 123
        explode: true
        in: path
        name: user_id
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      - description: The access level identifier.
        example: 123
        explode: true
        in: query
        name: access_level_id
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          description: The user was successfully removed from the resource.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - resources
  /groups:
    get:
      description: Returns a list of groups for your organization.
      operationId: getGroups
      parameters:
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsResponse'
          description: One page of groups that exist for your custom app.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
  /groups/{group_id}:
    get:
      description: Returns a group for your organization.
      operationId: getGroup
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'
          description: The group that exists for your custom app.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
  /groups/{group_id}/member-groups:
    post:
      description: Add a member group to the specified group
      operationId: addGroupMemberGroup
      parameters:
        - description: The ID of the parent group
          example: f454d283-ca67-4a8a-bdbb-df212eca5353
          in: path
          explode: true
          name: group_id
          required: true
          schema:
            type: string
          style: simple
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                group_id:
                  description: The ID of the member group to add.
                  example: f454d283-ca67-4a8a-bdbb-df212eca5353
                  type: string
                app_id:
                  description: The identifier of your custom app.
                  example: my-custom-app
                  type: string
              required:
                - group_id
                - app_id
      responses:
        "200":
          description: The group was successfully added to the group.
        "401":
          $ref: "#/components/responses/InvalidRequestSignatureError"
        "404":
          $ref: "#/components/responses/EntityNotFoundError"
        default:
          $ref: "#/components/responses/UnexpectedError"
      security:
        - BearerAuth: []
      tags:
        - groups
  /groups/{group_id}/member-groups/{member_group_id}:
    delete:
      description: Remove a member group from the specified group
      operationId: removeGroupMemberGroup
      parameters:
        - description: The ID of the parent group
          example: f454d283-ca67-4a8a-bdbb-df212eca5353
          in: path
          explode: true
          name: group_id
          required: true
          schema:
            type: string
          style: simple
        - description: The ID of the member group to remove.
          example: f454d283-ca67-4a8a-bdbb-df212eca5353
          in: path
          explode: true
          name: member_group_id
          required: true
          schema:
            type: string
          style: simple
        - description: The identifier of your custom app.
          example: my-custom-app
          explode: true
          in: query
          name: app_id
          required: true
          schema:
            type: string
          style: form
      responses:
        "200":
          description: The member group was successfully removed from the group
        "401":
          $ref: "#/components/responses/InvalidRequestSignatureError"
        "404":
          $ref: "#/components/responses/EntityNotFoundError"
        default:
          $ref: "#/components/responses/UnexpectedError"
      security:
        - BearerAuth: []
      tags:
        - groups
  /groups/{group_id}/users:
    get:
      description: Returns a list of users for the specified group.
      operationId: getGroupUsers
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupUsersResponse'
          description: One page of users that exist for your custom app.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
    post:
      description: Adds a user to the specified group.
      operationId: addGroupUser
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addGroupUser_request'
        required: true
      responses:
        "200":
          description: The user was successfully added to the group.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
  /groups/{group_id}/users/{user_id}:
    delete:
      description: Removes a user from the specified group.
      operationId: removeGroupUser
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the user.
        example: 123
        explode: true
        in: path
        name: user_id
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          description: The user was successfully removed from the group.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
  /groups/{group_id}/resources:
    get:
      description: Returns a list of resources for the specified group.
      operationId: getGroupResources
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      - description: The pagination cursor value.
        example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResourcesResponse'
          description: One page of resources that exist for your custom app.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
    post:
      description: Adds a resource to the specified group.
      operationId: addGroupResource
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addGroupResource_request'
        required: true
      responses:
        "200":
          description: The resource was successfully added to the group.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
  /groups/{group_id}/resources/{resource_id}:
    delete:
      description: Removes a resource from the specified group.
      operationId: removeGroupResource
      parameters:
      - description: The id of the group.
        example: f454d283-ca67-4a8a-bdbb-df212eca5353
        explode: true
        in: path
        name: group_id
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the resource.
        example: 123
        explode: true
        in: path
        name: resource_id
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of your custom app.
        example: my-custom-app
        explode: true
        in: query
        name: app_id
        required: true
        schema:
          type: string
        style: form
      - description: The access level identifier.
        example: 123
        explode: true
        in: query
        name: access_level_id
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          description: The resource was successfully removed from the group.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request signature is invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The entity was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: An unexpected error occurred.
      security:
      - BearerAuth: []
      tags:
      - groups
components:
  responses:
...
    InvalidRequestSignatureError:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: The request signature is invalid.
...
  schemas:
    Error:
      example:
        code: 0
        message: message
      properties:
        code:
          type: integer
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    ResourceResponse:
      example:
        resource:
        - id: f454d283-ca67-4a8a-bdbb-df212eca5353
          name: Admin
          description: This resource represents the admin role.
      properties:
        resource:
          $ref: '#/components/schemas/Resource'
      required:
      - resource
      type: object
    UsersResponse:
      example:
        next_cursor: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        users:
        - id: f454d283-ca67-4a8a-bdbb-df212eca5353
          email: john.doe@mycompnay.com
      properties:
        next_cursor:
          description: The cursor with which to continue pagination if additional
            result pages exist.
          example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
          nullable: true
          type: string
        users:
          items:
            $ref: '#/components/schemas/User'
          type: array
      required:
      - users
      type: object
    ResourcesResponse:
      example:
        next_cursor: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
        resources:
        - id: f454d283-ca67-4a8a-bdbb-df212eca5353
          name: Admin
          description: This resource represents the admin role.
      properties:
        next_cursor:
          description: The cursor with which to continue pagination if additional
            result pages exist.
          example: cD0yMDIxLTAxLTA2KzAzJTNBMjQlM0E1My40MzQzMjYlMkIwMCUzQTAw
          nullable: true
          type: string
        resources:
          items:
            $ref: '#/components/schemas/Resource'
          type: array
      required:
      - resources
      type: object
...